src/main/webapp/app/entities/employee-goals-review/delete/employee-goals-review-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { EmployeeGoalsReviewService } from '../service/employee-goals-review.service';

import { EmployeeGoalsReviewDeleteDialogComponent } from './employee-goals-review-delete-dialog.component';

describe('EmployeeGoalsReview Management Delete Component', () => {
  let comp: EmployeeGoalsReviewDeleteDialogComponent;
  let fixture: ComponentFixture<EmployeeGoalsReviewDeleteDialogComponent>;
  let service: EmployeeGoalsReviewService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [EmployeeGoalsReviewDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(EmployeeGoalsReviewDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(EmployeeGoalsReviewDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(EmployeeGoalsReviewService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
